+++ CHANGES.md
@@ -2,7 +2,9 @@ GH::watch changelog
 ===================
 
 ## v1.22 (ToDo)
+- \#94 added app setting option to enable background notification check over WiFi only to save mobile data
 - \#92 added longer period options (6h, 12h, 24h) to "New notifications check period" 
+- \#93 changed default value of "New notifications check period" preference from 5 minutes to 1 hour to lover network bandwidth
 
 ## v1.21 (Oct 28th, 2017)
 - \#46 update to the Material design